% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpCNV.R
\name{cmpCNV}
\alias{cmpCNV}
\title{Comparison of CNVs between two samples}
\usage{
cmpCNV(sample1, sample2, default_cn = 2)
}
\arguments{
\item{sample1}{A GRanges object}

\item{sample2}{A GRanges object}

\item{default_cn}{numeric value(The copy number of the uncovered fragment in sample B for sample A ). Default: \code{2}.}
}
\value{
A GRanges object
}
\description{
Comparison of CNVs between two samples
}
\examples{
gr1_file <- system.file("extdata", "test1.txt", package = "CNVProcessor")
if (file.exists(gr1_file)) {
  gr1 <- loadCNVtoGranges(gr1_file, start = "start", end = "end", cn = "cn", seqnames = "seqnames")
}
head(gr1)
gr2_file <- system.file("extdata", "test2.txt", package = "CNVProcessor")
if (file.exists(gr2_file)) {
  gr2 <- loadCNVtoGranges(gr2_file, start = "start", end = "end", cn = "cn", seqnames = "seqnames")
}
head(gr2)
gr_merge <- cmpCNV(sample1=gr1,sample2=gr2)
head(gr_merge)
}
