% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpCNVplot.R
\name{cmpCNVplot}
\alias{cmpCNVplot}
\title{CNV plot}
\usage{
cmpCNVplot(
  data,
  genome_version = c("hg19", "hg38"),
  chroms = paste0("chr", c(1:22, "X", "Y")),
  fc_col = "FC",
  amp_color = "red",
  del_color = "blue",
  title = "",
  normal = 0,
  show_chrom_labels = TRUE,
  xlab = "Chromosome",
  ylab = "CNV Difference (Treated - Control)"
)
}
\arguments{
\item{data}{A GRanges object}

\item{genome_version}{Genome version to use for chromosome lengths. Options are \code{"hg19"} or \code{"hg38"}. Ignored if \code{seqlengths} is provided.}

\item{chroms}{The chromosome that needs to be plotted,Default: chr1-chr22,chrX,chrY}

\item{fc_col}{The colname the CN. Usually the fold Change column of the comparison results}

\item{amp_color}{The color of the amplified area, Default: red}

\item{del_color}{The color of the deleted area, Default: blue}

\item{title}{The title of the diagram}

\item{normal}{A numerical value that defines the number of normal chromosomal copies. Default: 0 (the difference is 0 in comparison)}

\item{show_chrom_labels}{A logical value that displays the chromosome tag or not}

\item{xlab}{The X-axis title}

\item{ylab}{The Y-axis title}
}
\value{
Diagram of chromosomal differences
}
\description{
It is mainly used to show the comparison results of samples CNVs.
}
\examples{
gr1_file <- system.file("extdata", "test1.txt", package = "CNVProcessor")
if (file.exists(gr1_file)) {
  gr1 <- loadCNVtoGranges(gr1_file, start = "start", end = "end", cn = "cn", seqnames = "seqnames")
}
gr2_file <- system.file("extdata", "test2.txt", package = "CNVProcessor")
if (file.exists(gr2_file)) {
  gr2 <- loadCNVtoGranges(gr2_file, start = "start", end = "end", cn = "cn", seqnames = "seqnames")
}
gr_merge <- cmpCNV(sample1=gr1,sample2=gr2)
p <- cmpCNVplot(data=gr_merge,genome_version = "hg19",fc_col = "FC")
}
