% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCNV2GR.R
\name{loadCNVtoGranges}
\alias{loadCNVtoGranges}
\title{Data Input}
\usage{
loadCNVtoGranges(
  cnvfile,
  start,
  end,
  cn,
  seqnames,
  header = TRUE,
  sep = "\\t",
  ...
)
}
\arguments{
\item{cnvfile}{The path where the file is located and the name of the file}

\item{start}{The column name of the start position}

\item{end}{The column name of the start position}

\item{cn}{The column name of the copy number}

\item{seqnames}{The column name of the chromosome}

\item{header}{a logical value indicating whether the file contains the names of the variables as its first line.Equated \code{\link[utils]{read.table}}}

\item{sep}{the field separator character. Values on each line of the file are separated by this character. Equated \code{\link[utils]{read.table}}}

\item{...}{}
}
\value{
A GRanges object with "cn"(copy number) as metadata
}
\description{
Load a table file with chromosome number, start position, end position, and copy number and convert it into a GRanges object
}
\examples{
 # Load sample data
cnv_file <- system.file("extdata", "test1.txt", package = "CNVProcessor")
if (file.exists(cnv_file)) {
  gr <- loadCNVtoGranges(cnv_file, start = "start", end = "end", cn = "cn", seqnames = "seqnames")
}
head(gr)
}
