% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mCNVplot.R
\name{mCNVplot}
\alias{mCNVplot}
\title{CNV plot}
\usage{
mCNVplot(
  data_list,
  genome_version = c("hg19", "hg38"),
  chroms = paste0("chr", c(1:22, "X", "Y")),
  cn_col = "cn",
  amp_color = "red",
  nor_color = "yellow",
  del_color = "blue",
  title = "",
  normal = 2,
  xlab = "Chromosome",
  ylab = "CNV Difference",
  show_chrom_labels = TRUE,
  sample_names = NULL
)
}
\arguments{
\item{data_list}{A list of GRanges object}

\item{genome_version}{Genome version to use for chromosome lengths. Options are \code{"hg19"} or \code{"hg38"}. Ignored if \code{seqlengths} is provided.}

\item{chroms}{The chromosome that needs to be plotted,Default: chr1-chr22,chrX,chrY}

\item{cn_col}{The colname the CN. Usually the fold Change column of the comparison results}

\item{amp_color}{The color of the amplified area, Default: red}

\item{nor_color}{The color of the normal area, Default: yellow}

\item{del_color}{The color of the deleted area, Default: blue}

\item{title}{The title of the diagram}

\item{normal}{A numerical value that defines the number of normal chromosomal copies. Default: 2}

\item{xlab}{The X-axis title}

\item{ylab}{The Y-axis title}

\item{show_chrom_labels}{A logical value that displays the chromosome tag or not}

\item{sample_names}{A string vector containing the sample names}
}
\value{
Diagram of the CNV of multiple samples.
}
\description{
Draw the CNV of multiple samples.
}
\examples{
gr1_file <- system.file("extdata", "test1.txt", package = "CNVProcessor")
if (file.exists(gr1_file)) {
  gr1 <- loadCNVtoGranges(gr1_file, start = "start", end = "end", cn = "cn", seqnames = "seqnames")
}
gr2_file <- system.file("extdata", "test2.txt", package = "CNVProcessor")
if (file.exists(gr2_file)) {
  gr2 <- loadCNVtoGranges(gr2_file, start = "start", end = "end", cn = "cn", seqnames = "seqnames")
}
gr3_file <- system.file("extdata", "test3.txt", package = "CNVProcessor")
if (file.exists(gr3_file)) {
  gr3 <- loadCNVtoGranges(gr3_file, start = "start", end = "end", cn = "cn", seqnames = "seqnames")
}
p <- mCNVplot(data_list=list(gr1=gr1,gr2=gr2,gr3=gr3),genome_version = "hg19",cn_col = "cn")
}
