% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFullCNV.R
\name{mergeFullCNV}
\alias{mergeFullCNV}
\title{Merge CNV region of samples}
\usage{
mergeFullCNV(
  sample_list,
  genome_version = c("hg19", "hg38"),
  default_cn = 2,
  seqlengths = NULL
)
}
\arguments{
\item{sample_list}{A list of GRanges objects}

\item{genome_version}{Genome version to use for chromosome lengths. Options are \code{"hg19"} or \code{"hg38"}.Ignored if \code{seqlengths} is provided.}

\item{default_cn}{numeric value(The copy number of the uncovered fragment).  Default: \code{2}.}

\item{seqlengths}{Optional named numeric vector of chromosome lengths (e.g., \code{c(chr1=249250621, ...)}).
If \code{NULL}, chromosome lengths are automatically loaded from the specified \code{genome_version}.}
}
\value{
A GRanges objects containing CNV information from multiple samples in the same region (but complete chromosomes)
}
\description{
Segment the CNV regions of multiple samples according to their loci, while populating the uncovered genomic fragments
}
\examples{
gr1_file <- system.file("extdata", "test1.txt", package = "CNVProcessor")
if (file.exists(gr1_file)) {
  gr1 <- loadCNVtoGranges(gr1_file, start = "start", end = "end", cn = "cn", seqnames = "seqnames")
}
head(gr1)
gr2_file <- system.file("extdata", "test2.txt", package = "CNVProcessor")
if (file.exists(gr2_file)) {
  gr2 <- loadCNVtoGranges(gr2_file, start = "start", end = "end", cn = "cn", seqnames = "seqnames")
}
head(gr2)
gr_merge <- mergeFullCNV(sample_list=list(gr1=gr1,gr2=gr2),genome_version ="hg19")
head(gr_merge)
}
