% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeRawCNV.R
\name{mergeRawCNV}
\alias{mergeRawCNV}
\title{Merge CNV region of samples}
\usage{
mergeRawCNV(sample_list, default_cn = 2)
}
\arguments{
\item{sample_list}{A list of GRanges objects}

\item{default_cn}{numeric value(The copy number of the uncovered fragment in sample B for sample A ). Default: \code{2}.}
}
\value{
A GRanges object containing CNV information for multiple samples in the same region
}
\description{
Segment the CNV regions of multiple samples according to their loci, but do not populate the uncovered genomic fragments
}
\examples{
gr1_file <- system.file("extdata", "test1.txt", package = "CNVProcessor")
if (file.exists(gr1_file)) {
  gr1 <- loadCNVtoGranges(gr1_file, start = "start", end = "end", cn = "cn", seqnames = "seqnames")
}
head(gr1)
gr2_file <- system.file("extdata", "test2.txt", package = "CNVProcessor")
if (file.exists(gr2_file)) {
  gr2 <- loadCNVtoGranges(gr2_file, start = "start", end = "end", cn = "cn", seqnames = "seqnames")
}
head(gr2)
gr_merge <- mergeRawCNV(sample_list=list(gr1=gr1,gr2=gr2))
head(gr_merge)
}
