% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCNVplot.R
\name{sCNVplot}
\alias{sCNVplot}
\title{CNV plot}
\usage{
sCNVplot(
  data,
  genome_version = c("hg19", "hg38"),
  chroms = paste0("chr", c(1:22, "X", "Y")),
  cn_col = "cn",
  amp_color = "red",
  nor_color = "yellow",
  del_color = "blue",
  title = "",
  normal = 2,
  xlab = "Chromosome",
  ylab = "CNV Difference (Treated - Control)",
  show_chrom_labels = TRUE
)
}
\arguments{
\item{data}{A GRanges object}

\item{genome_version}{Genome version to use for chromosome lengths. Options are \code{"hg19"} or \code{"hg38"}. Ignored if \code{seqlengths} is provided.}

\item{chroms}{The chromosome that needs to be plotted,Default: chr1-chr22,chrX,chrY}

\item{cn_col}{The colname the CN. Usually the fold Change column of the comparison results}

\item{amp_color}{The color of the amplified area, Default: red}

\item{nor_color}{The color of the normal area, Default: yellow}

\item{del_color}{The color of the deleted area, Default: blue}

\item{title}{The title of the diagram}

\item{normal}{A numerical value that defines the number of normal chromosomal copies. Default: 2}

\item{xlab}{The X-axis title}

\item{ylab}{The Y-axis title}

\item{show_chrom_labels}{A logical value that displays the chromosome tag or not}
}
\value{
Diagram of the CNV of a single sample.
}
\description{
Draw the CNV of a single sample.
}
\examples{
cnv_file <- system.file("extdata", "test1.txt", package = "CNVProcessor")
if (file.exists(cnv_file)) {
  gr <- loadCNVtoGranges(cnv_file, start = "start", end = "end", cn = "cn", seqnames = "seqnames")
}
p <- sCNVplot(data=gr,genome_version = "hg19",title = "",cn_col = "cn",normal = 2,chroms = paste0("chr", 1:5),ylab = "CNV",show_chrom_labels = T)
}
